extract new relic integration into a plugin
