+++ Gemfile
@@ -88,7 +88,6 @@ group :no_preload do
   gem 'active_hash', '~> 1.0'
   gem 'ansible'
   gem 'github-markdown', '~> 0.6.3'
-  gem 'newrelic_api'
   gem 'activeresource'
   gem 'coderay', '~> 1.1.0'
   gem 'dogapi', '~> 1.9'
