+++ app/controllers/concerns/stage_permitted_params.rb
@@ -1,20 +1,25 @@
 module StagePermittedParams
   def stage_permitted_params
-    @stage_permitted_params ||=
-      ([ :name, :command, :confirm, :permalink, :dashboard,
-        :production,
-        :notify_email_address,
-        :deploy_on_release,
-        :datadog_tags,
-        :datadog_monitor_ids,
-        :update_github_pull_requests,
-        :email_committers_on_automated_deploy_failure,
-        :static_emails_on_automated_deploy_failure,
-        :use_github_deployment_api,
-        :no_code_deployed,
+    @stage_permitted_params ||= ([
+      :name,
+      :command,
+      :confirm,
+      :permalink,
+      :dashboard,
+      :production,
+      :notify_email_address,
+      :deploy_on_release,
+      :datadog_tags,
+      :datadog_monitor_ids,
+      :update_github_pull_requests,
+      :email_committers_on_automated_deploy_failure,
+      :static_emails_on_automated_deploy_failure,
+      :use_github_deployment_api,
+      :no_code_deployed,
+      {
         deploy_group_ids: [],
         command_ids: [],
-        new_relic_applications_attributes: [:id, :name, :_destroy]
-      ] + Samson::Hooks.fire(:stage_permitted_params)).freeze
+      }
+    ] + Samson::Hooks.fire(:stage_permitted_params)).freeze
   end
 end
