+++ app/models/stage.rb
@@ -8,7 +8,6 @@ class Stage < ActiveRecord::Base
 
   has_many :deploys, dependent: :destroy
   has_many :webhooks, dependent: :destroy
-  has_many :new_relic_applications
 
   has_one :lock
 
@@ -22,7 +21,6 @@ class Stage < ActiveRecord::Base
   default_scope { order(:order) }
 
   validates :name, presence: true, uniqueness: { scope: [:project, :deleted_at] }
-  accepts_nested_attributes_for :new_relic_applications, allow_destroy: true, reject_if: :no_newrelic_name?
 
   before_create :ensure_ordering
 
@@ -35,7 +33,6 @@ class Stage < ActiveRecord::Base
   def self.build_clone(old_stage)
     new(old_stage.attributes.except("id")).tap do |new_stage|
       Samson::Hooks.fire(:stage_clone, old_stage, new_stage)
-      new_stage.new_relic_applications.build(old_stage.new_relic_applications.map { |app| app.attributes.except("id", "updated_at", "created_at") })
       new_stage.command_ids = old_stage.command_ids
     end
   end
@@ -171,10 +168,6 @@ class Stage < ActiveRecord::Base
   end
 
   private
-
-  def no_newrelic_name?(newrelic_attrs)
-    newrelic_attrs['name'].blank?
-  end
 
   def permalink_base
     name
