+++ lib/samson/hooks.rb
@@ -7,6 +7,8 @@ module Samson
       :stage_form,
       :project_form,
       :deploys_header,
+      :deploy_tab_nav,
+      :deploy_tab_body,
       :deploy_view,
       :deploy_form, # for external plugin, so they can add extra form fields
       :admin_menu,
