+++ plugins/new_relic/app/assets/javascripts/new_relic/application.js
