+++ plugins/new_relic/app/helpers/new_relic_helper.rb
@@ -1,0 +1,5 @@
+module NewRelicHelper
+  def newrelic_enabled_for_deploy?
+    NewRelicApi.api_key.present? && @deploy.stage.new_relic_applications.any?
+  end
+end
