+++ plugins/new_relic/app/models/new_relic_application.rb
