+++ plugins/new_relic/app/views/samson_new_relic/_deploy_tab_nav.html.erb
@@ -1,0 +1,10 @@
+<% if newrelic_enabled_for_deploy? %>
+  <li>
+    <a href="#newrelic" id="newrelic-tab"
+       data-enabled="<%= @deploy.stage.new_relic_applications.any? %>"
+       data-url="<%= new_relic_project_stage_path(@project, @deploy.stage, format: :json) %>"
+       data-interval="60000" data-toggle="tab">
+      New Relic
+    </a>
+  </li>
+<% end %>
