+++ plugins/new_relic/app/views/samson_new_relic/_fields.html.erb
@@ -1,0 +1,26 @@
+<fieldset>
+  <legend>New Relic</legend>
+  <% if NewRelicApi.api_key %>
+    <% form.object.new_relic_applications.build %>
+
+    <p class="col-lg-offset-2">
+      Names of New Relic Applications to display during deploy
+    </p>
+    <%= form.fields_for :new_relic_applications do |app_fields| %>
+      <div class="form-group">
+        <div class="col-lg-6 col-lg-offset-2">
+          <%= app_fields.text_field :name, class: "form-control", placeholder: "Application name" %>
+        </div>
+
+        <% if app_fields.object.persisted? %>
+          <div class="col-lg-1 checkbox">
+            <%= app_fields.check_box :_destroy if app_fields.object.persisted? %>
+            <%= app_fields.label :_destroy, "Delete" %>
+          </div>
+        <% end %>
+      </div>
+    <% end %>
+  <% else %>
+    Set NEWRELIC_API_KEY to use NewRelic stats (disable the new_relic plugin to hide this warning).
+  <% end %>
+</fieldset>
