+++ plugins/new_relic/config/routes.rb
@@ -1,0 +1,7 @@
+Samson::Application.routes.draw do
+  resources :projects do
+    resources :stages do
+      get :new_relic, to: 'new_relic#show'
+    end
+  end
+end
