+++ plugins/new_relic/db/migrate/20140131005054_add_stage_new_relic_applications.rb
