+++ plugins/new_relic/samson_new_relic.gemspec
@@ -1,0 +1,6 @@
+Gem::Specification.new "samson_new_relic", "0.0.0" do |s|
+  s.summary = "Samson NewRelic integration"
+  s.authors = ["Michael Grosser"]
+  s.email = "mgrosser@zendesk.com"
+  s.add_runtime_dependency "newrelic_api"
+end
