+++ plugins/new_relic/test/controllers/new_relic_controller_test.rb
@@ -3,46 +3,41 @@ require_relative '../test_helper'
 SingleCov.covered!
 
 describe NewRelicController do
-  before do
-    NewRelicApi.api_key = 'hello'
-  end
+  before { NewRelicApi.api_key = 'hello' }
+  after { NewRelicApi.api_key = nil }
 
   as_a_viewer do
-    unauthorized :get, :show, project_id: :foo, id: 1
+    unauthorized :get, :show, project_id: :foo, stage_id: 1
   end
 
   as_a_project_deployer do
     describe "#show" do
       it "requires a project" do
         assert_raises(ActiveRecord::RecordNotFound) do
-          get :show, project_id: 123123, id: 123123
+          get :show, project_id: 123123, stage_id: stages(:test_staging)
         end
       end
 
       it "requires a stage" do
         assert_raises(ActiveRecord::RecordNotFound) do
-          get :show, project_id: projects(:test), id: 123123
+          get :show, project_id: projects(:test), stage_id: 123123
         end
       end
 
       it "requires a NewReclic api key" do
-        begin
-          old, NewRelicApi.api_key = NewRelicApi.api_key, ""
-          get :show, project_id: projects(:test), id: 123123
-        ensure
-          NewRelicApi.api_key = old
-        end
+        NewRelicApi.api_key = ""
+        get :show, project_id: projects(:test), stage_id: stages(:test_staging)
         assert_response :precondition_failed
       end
 
       describe "success" do
         before do
-          NewRelicHelper.expects(:metrics).
+          SamsonNewRelic::Api.expects(:metrics).
             with([new_relic_applications(:production).name], initial).
             returns('test_project' => true)
 
           get :show, project_id: projects(:test),
-            id: stages(:test_staging).id,
+            stage_id: stages(:test_staging),
             initial: initial ? 'true' : nil
         end
