+++ plugins/new_relic/test/models/new_relic_application_test.rb
