+++ plugins/new_relic/test/test_helper.rb
@@ -1,0 +1,1 @@
+require_relative '../../../test/test_helper'
